[BREAKING] Handle multiple credentials per user properly.
A user can have multiple credentials registered. The code didn't handle this case at all.
verifyAuthenticationResponse only took a single CredentialEntry;  whilst the client can return
any credential in the allowCredentials list.

We fix this by passing a list of CredentialEntry's.

Furthermore we stop storing the CredentialOptions. Instead we only store the state that is neccessery
for the registration and authentication ceremonies.  This makes the code a bit easier to reason about
as we only use a few fields from CredentialOptions.

Next due to the code not handling multiple credentials, the code for handling the userHandle field was non-sensical.
When userHandle is set; we should do a lookup of the credential provided using the userHandle.
This is in the spec (was even in the commments) but was not implemented.
This is implemented now.

This means Discoverable Credential flow (as popularized by Passkeys) is implementable now

TODO:

Fix Main.hs
Fix tests
Add more tests

What was implemented before does not look correct at all.